requirements.txt
mss>=9.0.0
opencv-python>=4.8.0
numpy>=1.24.0
pyautogui>=0.9.54
Pillow>=10.0.0
keyboard>=0.13.5
pystray>=0.19.5
